+++ src/VEM.jl
@@ -47,6 +47,7 @@ export mesh_scalar
 export cells_per_axis
 export nodes_per_axis
 export node_spacing
+export mesh_bounds
 export mesh_nodes
 
 include("blob/redistribution_kernel.jl")
@@ -54,6 +55,9 @@ export AbstractRedistributionKernel
 export M4Prime
 export redistribution_weight
 export support_radius
+
+include("blob/redistribution.jl")
+export redistribute
 
 include("precompile.jl")